Fixes #5 Only bind to QUALITY_SELECTED once
As things turn out, the middleware constructor is called every time `setSource`
is called [1]. This was causing a new listener to get bound to
`QUALITY_SELECTED` on each change by the quality selector. :( This change makes
it so the listener is only bound on the initial creation of the player.

[1] https://github.com/videojs/video.js/blob/03529163b677001f2ca994608d2026c55b11b73d/src/js/tech/middleware.js#L66