feat: Support toggling the select icon with the source label (#28)
User options can now be passed to the plugin.  Options are passed upon initialization and use the following syntax:

require('@silvermine/videojs-quality-selector')(videojs, options);

If you're not using require, you can pass options using a window config:

<script>
   window.SILVERMINE_VIDEOJS_QUALITY_SELECTOR_CONFIG = {
      showQualitySelectionLabelInControlBar: true,
   };
</script>
<script src="path/to/silvermine-videojs-quality-selector.js"></script>

The options are sent from the constructor to `/components/QualitySelector.js`.

Currently only one option is available:

showQualitySelectionLabelInControlBar: (default: `false`) - Enabling this flag will swap the selector
icon with the label of the currently selected source.

More options should be possible by adding them to the object.

A class is added to the Menu Item (Has class vjs-icon-placeholder) when the flag is disabled to include the icon.  When enabled, basic HTML is simply included into the element.

Refs #28